Fix #1786 Refactor slack app admin.py into separate files (#1935)
* refactor slack app admin.py

* Fixed duplication

* used mixin base fields

* Fix SonarQube warnings

* removed mixins

* Update code

---------
