+++ backend/apps/slack/admin/__init__.py
@@ -1,0 +1,7 @@
+"""Slack app admin."""
+
+from .conversation import ConversationAdmin
+from .event import EventAdmin
+from .member import MemberAdmin
+from .message import MessageAdmin
+from .workspace import WorkspaceAdmin
